Updates SDK to version 1.5.14
Updates the SDK to the latest version, 1.5.14.

This change includes updating the version number in the
build.gradle file, the BuildInfo.kt file, and the pom.xml file.
It also includes the addition of the new aar and pom files to the
repository, as well as a new release process document and a new
cleanup script.